Save all .npy files from .tif files
